move sorting functions to new file
